+++ src/Utility/BackupImport.php
@@ -30,7 +30,7 @@ class BackupImport
      * Driver containing all methods to export/import database backups according to the connection
      * @var \DatabaseBackup\Driver\Driver
      */
-    protected Driver $Driver;
+    public Driver $Driver;
 
     /**
      * Filename where to import the database
@@ -68,12 +68,18 @@ class BackupImport
     }
 
     /**
-     * Imports the database
-     * @return string Filename path
+     * Imports the database.
+     *
+     * When importing, this method will trigger these events (implemented by the driver instance):
+     *  - `Backup.beforeImport`: will be triggered before import;
+     *  - `Backup.afterImport`: will be triggered after import.
+     * @return string|false Filename path on success or `false` if the `Backup.beforeImport` event is stopped
      * @throws \ErrorException|\ReflectionException
+     * @see \DatabaseBackup\Driver\Driver::afterImport()
+     * @see \DatabaseBackup\Driver\Driver::beforeImport()
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupImport-utility#import
      */
-    public function import(): string
+    public function import()
     {
         Exceptionist::isTrue(!empty($this->filename), __d('database_backup', 'You must first set the filename'));
 
@@ -81,7 +87,17 @@ class BackupImport
         $filename = $this->filename;
         unset($this->filename);
 
+        //Dispatches the `Backup.beforeImport` event implemented by the driver
+        $BeforeImport = $this->Driver->dispatchEvent('Backup.beforeImport');
+        if ($BeforeImport->isStopped()) {
+            return false;
+        }
+
+        //Imports
         $this->Driver->import($filename);
+
+        //Dispatches the `Backup.afterImport` event implemented by the driver
+        $this->Driver->dispatchEvent('Backup.afterImport');
 
         return $filename;
     }